+++ changelog.md
@@ -1,4 +1,5 @@
 13.03.2025 - Repository created, OpenAI test request added\
 15.03.2025 - Cleaned API key from repository\
-16.03.2025 - Added Playwright and screenshot functionality
-22.03.2025 - Added database
+16.03.2025 - Added Playwright and screenshot functionality\
+22.03.2025 - Added database\
+22.03.2025 - Screenshots from heatmap are saved to database as base64